Added Cube Getting Number on Faces Script, oncollision change or add number on dice
At start value of Text added on the cube can be changed.When play is pressed then number plates whose number are already on the dice are deactivated for now and if any face face has a number on it then the the face opposite to it gets a number as per die look.
if die collides with number plate then if the face that collide does not have any number then number is added and number plate is deactivated and opposite side face gets a number also. new number plates are tracked in the code.
else if the face that touched the number has number on it then older number plates are activated and the number on the number plate is added to the face and  a number is added to the opposite face also.
Also the newer number plates are tracked in the code as well as deactivated so as to not cause any issues.